Converted `/api/generateTokenIdea` to be framework agnostic
