+++ libs/model/package.json
@@ -46,6 +46,7 @@
     "moment": "^2.23.0",
     "node-fetch": "2",
     "node-object-hash": "^3.0.0",
+    "openai": "^4.0.0",
     "pg": "^8.11.3",
     "quill-delta-to-markdown": "^0.6.0",
     "sequelize": "^6.32.1",
