+++ libs/model/src/services/index.ts
@@ -1,4 +1,5 @@
 export * as commonProtocol from './commonProtocol';
+export * from './openai';
 export * from './snapshot';
 export * as stakeHelper from './stakeHelper';
 export * as tokenBalanceCache from './tokenBalanceCache';
