+++ libs/model/src/services/openai/index.ts
@@ -1,0 +1,1 @@
+export * from './generateTokenIdea';
